+++ imap_tools/__init__.py
@@ -1,7 +1,7 @@
-from .query import Q, AND, OR, NOT, Header, A, O, N, H
+from .query import AND, OR, NOT, Header, UidRange, A, O, N, H, U
 from .mailbox import BaseMailBox, MailBox, MailBoxUnencrypted
 from .message import MailMessage, Attachment, MailMessageFlags
 from .folder import MailBoxFolderManager, MailBoxFolderStatusOptions
 from .errors import *
 
-__version__ = '0.30.0'
+__version__ = '0.31.0'