+++ imap_tools/mailbox.py
@@ -1,9 +1,10 @@
+import sys
 import imaplib
 from email.errors import StartBoundaryNotFoundDefect, MultipartInvariantViolationDefect
 
 from .message import MailMessage, MailMessageFlags
 from .folder import MailBoxFolderManager
-from .utils import cleaned_uid_set, check_command_status, grouper
+from .utils import cleaned_uid_set, check_command_status, chunks
 from .errors import MailboxStarttlsError, MailboxLoginError, MailboxLogoutError, MailboxSearchError, \
     MailboxFetchError, MailboxExpungeError, MailboxDeleteError, MailboxCopyError, MailboxFlagError
 
@@ -17,6 +18,7 @@ class BaseMailBox:
     email_message_class = MailMessage
     folder_manager_class = MailBoxFolderManager
     with_headers_only_allowed_errors = (StartBoundaryNotFoundDefect, MultipartInvariantViolationDefect)
+    timeout_not_supported_error = 'timeout argument supported since python 3.9'
 
     def __init__(self):
         self.folder = None  # folder manager
@@ -63,7 +65,7 @@ class BaseMailBox:
             return
         fetch_result = self.box.fetch(','.join(message_nums), message_parts)
         check_command_status(fetch_result, MailboxFetchError)
-        for built_fetch_item in grouper((reversed if reverse else iter)(fetch_result[1]), 2):
+        for built_fetch_item in chunks((reversed if reverse else iter)(fetch_result[1]), 2):
             yield built_fetch_item
 
     def fetch(self, criteria: str or bytes = 'ALL', charset: str = 'US-ASCII', limit: int or slice = None,
@@ -183,33 +185,44 @@ class BaseMailBox:
 class MailBoxUnencrypted(BaseMailBox):
     """Working with the email box through IMAP4"""
 
-    def __init__(self, host='', port=143):
+    def __init__(self, host='', port=143, timeout=None):
         """
         :param host: host's name (default: localhost)
         :param port: port number
+        :param timeout: timeout in seconds for the connection attempt, since python 3.9
         """
+        if timeout and sys.version_info.minor < 9:
+            raise ValueError(self.timeout_not_supported_error)
         self._host = host
         self._port = port
+        self._timeout = timeout
         super().__init__()
 
     def _get_mailbox_client(self):
-        return imaplib.IMAP4(self._host, self._port)
+        if sys.version_info.minor < 9:
+            return imaplib.IMAP4(self._host, self._port)
+        else:
+            return imaplib.IMAP4(self._host, self._port, self._timeout)  # noqa
 
 
 class MailBox(BaseMailBox):
     """Working with the email box through IMAP4 over SSL connection"""
 
-    def __init__(self, host='', port=993, keyfile=None, certfile=None, ssl_context=None, starttls=False):
+    def __init__(self, host='', port=993, timeout=None, keyfile=None, certfile=None, ssl_context=None, starttls=False):
         """
         :param host: host's name (default: localhost)
         :param port: port number
+        :param timeout: timeout in seconds for the connection attempt, since python 3.9
         :param keyfile: PEM formatted file that contains your private key (deprecated)
         :param certfile: PEM formatted certificate chain file (deprecated)
         :param ssl_context: SSLContext object that contains your certificate chain and private key
         :param starttls: whether to use starttls
         """
+        if timeout and sys.version_info.minor < 9:
+            raise ValueError(self.timeout_not_supported_error)
         self._host = host
         self._port = port
+        self._timeout = timeout
         self._keyfile = keyfile
         self._certfile = certfile
         self._ssl_context = ssl_context
@@ -220,12 +233,19 @@ class MailBox(BaseMailBox):
         if self._starttls:
             if self._keyfile or self._certfile:
                 raise ValueError("starttls cannot be combined with keyfile neither with certfile.")
-            client = imaplib.IMAP4(self._host, self._port)
+            if sys.version_info.minor < 9:
+                client = imaplib.IMAP4(self._host, self._port)
+            else:
+                client = imaplib.IMAP4(self._host, self._port, self._timeout)  # noqa
             result = client.starttls(self._ssl_context)
             check_command_status(result, MailboxStarttlsError)
             return client
         else:
-            return imaplib.IMAP4_SSL(self._host, self._port, self._keyfile, self._certfile, self._ssl_context)
+            if sys.version_info.minor < 9:
+                return imaplib.IMAP4_SSL(self._host, self._port, self._keyfile, self._certfile, self._ssl_context)
+            else:
+                return imaplib.IMAP4_SSL(self._host, self._port, self._keyfile,  # noqa
+                                         self._certfile, self._ssl_context, self._timeout)
 
     def xoauth2(self, username: str, access_token: str, initial_folder: str = 'INBOX'):
         """Authenticate to account using OAuth 2.0 mechanism"""